requirements.txt
django-filter
Django==1.11.11
django-rest-framework==0.1.0
djangorestframework==3.7.7
psycopg2==2.7.4
pytz==2018.3
